Add fetchPlan action
Wrote the fetchPlan action which is used to generate an itinerary
a user searches a city and is directed to the show page, or
the user decides to generate a new plan for a given city.

The fetchPlan action makes use of the Axios AJAX call which fetches
an itinerary for any given city.