Refactor file tree
Created a plans-index directory and moved everything PlansIndex
related to that directory. Import file paths were also updated to reflect
these changes. The biggest change is the extraction of the grid of the
plan cards themselves to separate files. This was done so that I can
more easily do the Redux cycle for each individual plan card.